components/job_configuration/python/requirements.txt
requests==2.21.0
PyYAML==5.2
